refactor: v1.0 stabilization — dead code, duplicates, circular deps
Phase 1: Remove unused dependencies and dead code
- Remove 5 unused npm packages: uuid, tar-fs, docx, @types/uuid, @types/tar-fs
- Delete orphan AttachmentContentModal.svelte component
- Remove unused SystemPromptConfig interface, preserveScroll function, FULL_CONTENT_MAX_CHARS export

Phase 2: Consolidate duplicate utilities
- Merge client/server token estimation into unified $lib/utils/tokens.ts
- Extract cosineSimilarity to shared math.ts utility
- Remove 4 clip() wrapper functions, inline clipText() calls
- Centralize DAY_MS and RERANK_CONFIDENCE_MIN constants

Phase 3: Deduplicate type definitions
- Replace local TemporalMetadata with PersonaMemoryTemporalInfo import
- Delete orphan admin.ts client API module

Phase 4: Break circular dependencies
- Extract buildPersonaPromptContext to persona-memory/context-builder.ts
- Extract syncArtifactChunks to task-state/chunk-sync.ts
- Break persona-memory <-> honcho and task-state <-> knowledge/store cycles

No behavioral changes.
