+++ package.json
@@ -35,9 +35,7 @@
 		"@types/jszip": "^3.4.0",
 		"@types/marked": "^5.0.2",
 		"@types/pdfjs-dist": "^2.10.377",
-		"@types/tar-fs": "^2.0.4",
 		"@types/tar-stream": "^3.1.4",
-		"@types/uuid": "^10.0.0",
 		"dotenv": "^17.3.1",
 		"drizzle-kit": "^0.31.9",
 		"express": "^5.2.1",
@@ -59,7 +57,6 @@
 		"better-sqlite3": "^12.8.0",
 		"chart.js": "^4.5.1",
 		"dockerode": "^4.0.10",
-		"docx": "^9.6.1",
 		"drizzle-orm": "^0.45.2",
 		"exceljs": "^4.4.0",
 		"gray-matter": "^4.0.3",
@@ -73,9 +70,7 @@
 		"pptxgenjs": "^4.0.1",
 		"pptxviewjs": "^1.1.9",
 		"shiki": "^4.0.2",
-		"tar-fs": "^3.1.2",
 		"tar-stream": "^3.1.8",
-		"uuid": "^13.0.0",
 		"zod": "^4.1.12"
 	},
 	"type": "module"
