+++ src/lib/server/services/persona-memory.ts
@@ -32,17 +32,12 @@ import {
 import { listPersonaMemories } from './honcho';
 import { recordMemoryEvents } from './memory-events';
 import { queuePersonaClusterSemanticEmbeddingRefresh } from './semantic-embedding-refresh';
-import { shortlistSemanticMatchesBySubject } from './semantic-ranking';
 import { canUseContextSummarizer, requestStructuredControlModel } from './task-state';
 import { classifyMemoryBatch } from './task-state/control-model';
-import { shouldIncludePersonaMemoryInGeneratedContext } from '../utils/conversation-boundary-filter';
 import {
 	determineTeiWinningMode,
 	logTeiRetrievalSummary,
-	type SemanticShortlistDiagnostics,
-	type TeiRerankDiagnostics,
 } from './tei-observability';
-import { canUseTeiReranker, rerankItems } from './tei-reranker';
 import { scoreMatch } from './working-set';
 import type { HonchoPersonaMemoryRecord } from './honcho';
 // Classification helpers (extracted to sub-module)
@@ -65,7 +60,7 @@ export {
 
 // Temporal helpers (extracted to sub-module)
 import {
-    deriveTemporalMetadata,
+    derivePersonaMemoryTemporalInfo,
     deriveTopicStatus,
     getTemporalFreshness,
     hasResolvedTemporalCue,
@@ -73,20 +68,14 @@ import {
 } from './persona-memory/temporal';
 
 
-const DAY_MS = 86_400_000;
 const DREAM_MIN_CHANGES = 10;
 const DREAM_INTERVAL_MS = DAY_MS;
 const FULL_SWEEP_INTERVAL_MS = 7 * DAY_MS;
-const ACTIVE_PROMPT_LIMIT = 8;
-const DORMANT_PROMPT_LIMIT = 2;
-const PROMPT_TEXT_BUDGET = 1600;
-const PROMPT_REFRESH_THROTTLE_MS = 60_000;
-const PERSONA_SEMANTIC_SHORTLIST_LIMIT = 12;
-const PERSONA_RERANK_LIMIT = 8;
+
 const SEMANTIC_RECONCILE_MIN_CONFIDENCE = 72;
 const MAX_SEMANTIC_CANDIDATES = 4;
-const ensureClustersReadyInFlight = new Map<string, Promise<void>>();
-const promptRefreshTriggeredAt = new Map<string, number>();
+export const ensureClustersReadyInFlight = new Map<string, Promise<void>>();
+export const promptRefreshTriggeredAt = new Map<string, number>();
 const personaRuntimeEpochByUser = new Map<string, number>();
 
 type ClusterPlan = {
@@ -893,23 +882,11 @@ function clusterIdForKey(userId: string, key: string): string {
 	return `pmc_${hashKey(`${userId}:${key}`)}`;
 }
 
-function clip(value: string, maxLength: number): string {
-	return clipText(value, maxLength);
-}
 
 function parseJsonRecord(value: string | null): Record<string, unknown> {
 	return parseJsonRecordOrNull(value) ?? {};
 }
 
-type TemporalMetadata = {
-	kind: PersonaMemoryTemporalKind;
-	freshness: PersonaMemoryTemporalFreshness;
-	observedAt: number;
-	effectiveAt: number | null;
-	expiresAt: number | null;
-	relative: boolean;
-	resolved: boolean;
-};
 
 function parseNumberToken(value: string | undefined): number | null {
 	if (!value) return null;
@@ -979,7 +956,7 @@ function buildHistoricalTemporalText(text: string, observedAt: number): string {
 
 function temporalMetadataFromRecord(
 	metadata: Record<string, unknown> | null
-): TemporalMetadata | null {
+): PersonaMemoryTemporalInfo | null {
 	if (!metadata) return null;
 	const temporal = metadata.temporal;
 	if (!temporal || typeof temporal !== 'object' || Array.isArray(temporal)) return null;
@@ -1350,7 +1327,7 @@ function deriveCanonicalText(params: {
 			return right.createdAt - left.createdAt;
 		})[0];
 
-	return clip(representative?.content ?? '', 320);
+	return clipText(representative?.content ?? '', 320);
 }
 
 function deriveSemanticFingerprint(text: string): SemanticFingerprint {
@@ -1540,7 +1517,7 @@ async function applyTargetedSemanticSupersession(plans: ClusterPlan[]): Promise<
 					supersessionReason: 'semantic_replace',
 					semanticSupersessionConfidence: confidence,
 					semanticSupersessionRationale:
-						typeof response?.rationale === 'string' ? clip(response.rationale, 220) : null,
+						typeof response?.rationale === 'string' ? clipText(response.rationale, 220) : null,
 				};
 				candidate.state = 'archived';
 				candidate.decayAt = null;
@@ -2068,7 +2045,7 @@ async function dreamCluster(params: {
 				: params.defaultMemoryClass;
 
 		return {
-			canonicalText: clip(
+			canonicalText: clipText(
 				sanitizeDreamedCanonicalText({
 					canonicalText: response?.canonicalText,
 					defaultCanonicalText: params.defaultCanonicalText,
@@ -2457,7 +2434,7 @@ export async function syncPersonaMemoryClusters(params: {
 				normalizedMemoryClass === 'stable_preference'
 					? extractPreferenceSlotMetadata(dream.canonicalText)
 					: null;
-			const temporalMetadata = deriveTemporalMetadata({
+			const temporalMetadata = derivePersonaMemoryTemporalInfo({
 				canonicalText: dream.canonicalText,
 				records: group.records,
 				memoryClass: normalizedMemoryClass,
@@ -2755,303 +2732,6 @@ export async function getPersonaMemoryClusterConclusionIds(
 	return rows.map((row) => row.conclusionId);
 }
 
-async function buildPersonaQueryScoreMaps(params: {
-	userId: string;
-	query: string;
-	items: PersonaMemoryItem[];
-}): Promise<{
-	semanticScoreById: Map<string, number>;
-	rerankScoreById: Map<string, number>;
-	semanticDiagnostics: SemanticShortlistDiagnostics | null;
-	rerankDiagnostics: TeiRerankDiagnostics | null;
-}> {
-	const trimmedQuery = params.query.trim();
-	if (!trimmedQuery || params.items.length === 0) {
-		return {
-			semanticScoreById: new Map(),
-			rerankScoreById: new Map(),
-			semanticDiagnostics: null,
-			rerankDiagnostics: null,
-		};
-	}
-
-	let semanticDiagnostics: SemanticShortlistDiagnostics | null = null;
-	const semanticMatches =
-		(await shortlistSemanticMatchesBySubject({
-			userId: params.userId,
-			subjectType: 'persona_cluster',
-			query: trimmedQuery,
-			items: params.items,
-			getSubjectId: (item) => item.id,
-			limit: PERSONA_SEMANTIC_SHORTLIST_LIMIT,
-			onDiagnostics: (diagnostics) => {
-				semanticDiagnostics = diagnostics;
-			},
-		})) ?? [];
-	const semanticScoreById = new Map(
-		semanticMatches.map((match) => [match.subjectId, match.semanticScore])
-	);
-
-	let rerankScoreById = new Map<string, number>();
-	let rerankDiagnostics: TeiRerankDiagnostics | null = null;
-	if (canUseTeiReranker() && semanticMatches.length > 1) {
-		try {
-			const reranked = await rerankItems({
-				query: trimmedQuery,
-				items: semanticMatches.map((match) => match.item),
-				getText: (item) => item.rawCanonicalText ?? item.canonicalText,
-				maxTexts: PERSONA_RERANK_LIMIT,
-				onDiagnostics: (diagnostics) => {
-					rerankDiagnostics = diagnostics;
-				},
-			});
-
-			if (reranked && reranked.items.length > 0) {
-				rerankScoreById = new Map(
-					reranked.items.map((entry) => [entry.item.id, entry.score])
-				);
-			}
-		} catch (error) {
-			console.error('[PERSONA_MEMORY] Semantic reranker failed:', {
-				userId: params.userId,
-				error,
-			});
-		}
-	}
-
-	return {
-		semanticScoreById,
-		rerankScoreById,
-		semanticDiagnostics,
-		rerankDiagnostics,
-	};
-}
-
-function getPersonaLexicalMatchScore(item: PersonaMemoryItem, query: string): number {
-	return Math.max(
-		scoreMatch(query, item.rawCanonicalText ?? item.canonicalText),
-		item.topicKey ? scoreMatch(query, item.topicKey) : 0
-	);
-}
-
-function getPersonaQueryMatchScore(params: {
-	item: PersonaMemoryItem;
-	query: string;
-	semanticScoreById: Map<string, number>;
-	rerankScoreById: Map<string, number>;
-}): number {
-	const lexicalScore = getPersonaLexicalMatchScore(params.item, params.query);
-	const semanticScore = params.semanticScoreById.get(params.item.id) ?? 0;
-	const rerankScore = params.rerankScoreById.get(params.item.id) ?? 0;
-
-	return lexicalScore + semanticScore * 3 + rerankScore * 4;
-}
-
-export async function buildPersonaPromptContext(
-	userId: string,
-	query: string
-): Promise<string> {
-	const now = Date.now();
-	const lastRefreshAt = promptRefreshTriggeredAt.get(userId) ?? 0;
-	if (
-		!ensureClustersReadyInFlight.has(userId) &&
-		now - lastRefreshAt >= PROMPT_REFRESH_THROTTLE_MS
-	) {
-		promptRefreshTriggeredAt.set(userId, now);
-		void ensurePersonaMemoryClustersReady(userId, 'prompt_read').catch((error) => {
-			console.warn('[PERSONA_MEMORY] Background cluster refresh failed', {
-				userId,
-				reason: 'prompt_read',
-				error,
-			});
-		});
-	}
-
-	const items = (await listPersonaMemoryClusters(userId)).filter((item) => {
-		if (item.state === 'archived') return false;
-		if (item.topicStatus === 'historical') return false;
-		if (item.temporal?.freshness === 'expired' || item.temporal?.freshness === 'historical') {
-			return false;
-		}
-		return true;
-	});
-	if (items.length === 0) return '';
-
-	const { semanticScoreById, rerankScoreById, semanticDiagnostics, rerankDiagnostics } =
-		await buildPersonaQueryScoreMaps({
-		userId,
-		query,
-		items,
-	});
-
-	const isGeneratedDocumentRequest =
-		/\bgenerate_file\b/i.test(query) ||
-		(/\b(?:generate|create|write|draft|make)\b/i.test(query) &&
-			/\b(?:file|document|pdf|report|presentation|deck|slide|spreadsheet)\b/i.test(query));
-
-	const activeConstraints = items
-		.filter((item) => item.state === 'active' && item.activeConstraint)
-		.map((item) => ({
-			item,
-			matchScore: getPersonaQueryMatchScore({
-				item,
-				query,
-				semanticScoreById,
-				rerankScoreById,
-			}),
-			constraintRank:
-				item.activeConstraint && item.temporal?.freshness === 'active'
-					? 2
-					: item.activeConstraint && item.temporal?.freshness === 'stale'
-						? 1
-						: 0,
-		}))
-		.filter((entry) =>
-			shouldIncludePersonaMemoryInGeneratedContext({
-				memoryCanonicalText: entry.item.canonicalText,
-				currentQuery: query,
-				queryOverlap: entry.matchScore,
-				isGeneratedDocumentRequest,
-			})
-		)
-		.sort(
-			(left, right) =>
-				right.constraintRank - left.constraintRank ||
-				right.matchScore - left.matchScore ||
-				right.item.salienceScore - left.item.salienceScore ||
-				right.item.lastSeenAt - left.item.lastSeenAt
-		)
-		.slice(0, ACTIVE_PROMPT_LIMIT);
-	const active = items
-		.filter((item) => item.state === 'active' && !item.activeConstraint)
-		.map((item) => ({
-			item,
-			matchScore: getPersonaQueryMatchScore({
-				item,
-				query,
-				semanticScoreById,
-				rerankScoreById,
-			}),
-		}))
-		.filter((entry) =>
-			shouldIncludePersonaMemoryInGeneratedContext({
-				memoryCanonicalText: entry.item.canonicalText,
-				currentQuery: query,
-				queryOverlap: entry.matchScore,
-				isGeneratedDocumentRequest,
-			})
-		)
-		.sort(
-			(left, right) =>
-				right.matchScore - left.matchScore ||
-				right.item.salienceScore - left.item.salienceScore ||
-				right.item.lastSeenAt - left.item.lastSeenAt
-		)
-		.slice(0, ACTIVE_PROMPT_LIMIT);
-	const dormant = items
-		.filter((item) => item.state === 'dormant' && !item.activeConstraint)
-		.map((item) => ({
-			item,
-			matchScore: getPersonaQueryMatchScore({
-				item,
-				query,
-				semanticScoreById,
-				rerankScoreById,
-			}),
-		}))
-		.filter((entry) => entry.matchScore >= 0.1)
-		.filter((entry) =>
-			shouldIncludePersonaMemoryInGeneratedContext({
-				memoryCanonicalText: entry.item.canonicalText,
-				currentQuery: query,
-				queryOverlap: entry.matchScore,
-				isGeneratedDocumentRequest,
-			})
-		)
-		.sort(
-			(left, right) =>
-				right.matchScore - left.matchScore ||
-				Number(right.item.activeConstraint) - Number(left.item.activeConstraint) ||
-				right.item.salienceScore - left.item.salienceScore
-		)
-		.slice(0, DORMANT_PROMPT_LIMIT)
-		.map((entry) => entry.item);
-
-	const selected = Array.from(
-		new Map(
-			[
-				...activeConstraints.map((entry) => entry.item),
-				...active.map((entry) => entry.item),
-				...dormant,
-			].map((item) => [item.id, item])
-		).values()
-	);
-	if (selected.length === 0) return '';
-
-	const winningCandidates = [
-		...activeConstraints.map((entry) => ({
-			item: entry.item,
-			matchScore: entry.matchScore,
-		})),
-		...active.map((entry) => ({
-			item: entry.item,
-			matchScore: entry.matchScore,
-		})),
-		...dormant.map((item) => ({
-			item,
-			matchScore: getPersonaQueryMatchScore({
-				item,
-				query,
-				semanticScoreById,
-				rerankScoreById,
-			}),
-		})),
-	].sort(
-		(left, right) =>
-			right.matchScore - left.matchScore ||
-			right.item.salienceScore - left.item.salienceScore ||
-			right.item.lastSeenAt - left.item.lastSeenAt
-	);
-	const winningCandidate = winningCandidates[0] ?? null;
-	logTeiRetrievalSummary({
-		scope: 'persona_prompt',
-		userId,
-		queryLength: query.trim().length,
-		candidateCount: items.length,
-		semantic: semanticDiagnostics,
-		rerank: rerankDiagnostics,
-		winningMode: determineTeiWinningMode({
-			lexicalScore: winningCandidate
-				? getPersonaLexicalMatchScore(winningCandidate.item, query)
-				: 0,
-			semanticScore: winningCandidate
-				? semanticScoreById.get(winningCandidate.item.id) ?? 0
-				: 0,
-			rerankScore: winningCandidate
-				? rerankScoreById.get(winningCandidate.item.id) ?? 0
-				: 0,
-		}),
-		winnerId: winningCandidate?.item.id ?? null,
-		extra: {
-			selectedCount: selected.length,
-			activeConstraintCount: activeConstraints.length,
-			activeCount: active.length,
-			dormantCount: dormant.length,
-		},
-	});
-
-	const lines: string[] = [];
-	let used = 0;
-	for (const item of selected) {
-		const line = `- ${item.canonicalText}`;
-		used += line.length;
-		if (used > PROMPT_TEXT_BUDGET) break;
-		lines.push(line);
-	}
-
-	return lines.length > 0 ? lines.join('\n') : '';
-}
-
 export async function deletePersonaMemoryClustersForConclusionIds(
 	userId: string,
 	conclusionIds: string[]
@@ -3089,3 +2769,5 @@ export async function deleteAllPersonaMemoryStateForUser(userId: string): Promis
 		tx.delete(personaMemoryClusters).where(eq(personaMemoryClusters.userId, userId)).run();
 	});
 }
+
+export { buildPersonaPromptContext } from './context-builder';
