--- src/lib/server/services/persona-memory/_constants.ts
@@ -1,1 +1,0 @@
-export const DAY_MS = 86_400_000;
