+++ src/lib/server/utils/constants.ts
@@ -1,0 +1,5 @@
+/** Milliseconds per day */
+export const DAY_MS = 86_400_000;
+
+/** Minimum confidence threshold for reranker results to be considered valid */
+export const RERANK_CONFIDENCE_MIN = 64;
