+++ src/lib/server/utils/topic-shift-detector.ts
@@ -1,23 +1,4 @@
-export function cosineSimilarity(a: number[], b: number[]): number {
-	if (a.length === 0 || b.length === 0) return 0;
-
-	const dimensions = Math.min(a.length, b.length);
-	if (dimensions === 0) return 0;
-
-	let dot = 0;
-	let aNorm = 0;
-	let bNorm = 0;
-
-	for (let i = 0; i < dimensions; i++) {
-		dot += a[i] * b[i];
-		aNorm += a[i] * a[i];
-		bNorm += b[i] * b[i];
-	}
-
-	if (aNorm <= 0 || bNorm <= 0) return 0;
-
-	return dot / (Math.sqrt(aNorm) * Math.sqrt(bNorm));
-}
+import { cosineSimilarity } from '$lib/server/utils/math';
 
 export interface DetectTopicShiftParams {
 	currentMessageEmbedding: number[];
